Tokenizacion agregada, middleware para autenticar el token agregado

+++ controllers/usuarios.controller.js
@@ -1,6 +1,7 @@
 const prisma = require("../orm")
 const { isNumeric, isEmpty } = require("validator")
 const bcrypt = require("bcryptjs")
+const jwt = require("../services/jwt.service")
 
 module.exports = {
     get: async (req, res) => {
@@ -99,8 +100,8 @@ module.exports = {
             if (!bcrypt.compareSync(password, data.contrasena)) {
                 return res.json({ status: 409, message: "Contraseña invalida" })
             }
-            
-            return res.json({ status: 200, message: "Ingresaste" })
+
+            return res.json({ status: 200, message: "Ingresaste", token: jwt.token(data) })
 
         }).catch(error => {
             return res.json({ status: 400, message: "Server error" })
@@ -111,7 +112,7 @@ module.exports = {
     },
 
     updatePassword: async (req, res) => {
-        const id = req.body.id ?? ""
+        const id = req.params.id ?? ""
         const password = req.body.password ?? ""
 
         if (!isNumeric(id) || isEmpty(password)) {
@@ -139,7 +140,7 @@ module.exports = {
 
     update: async (req, res) => {
 
-        const id = req.body.id ?? ""
+        const id = req.params.id ?? ""
         const name = req.body.name ?? ""
         const usuario = req.body.usuario ?? ""
 
@@ -148,41 +149,77 @@ module.exports = {
             return res.json({ status: 400, message: "No pasó la validación" })
         }
 
-        await prisma.usuarios.update({
+        await prisma.usuarios.findFirst({
             where: {
                 id: parseInt(id)
-            },
-            data: {
-                nombre: name.toLowerCase(),
-                usuario: usuario.toLowerCase()
             }
-        }).then(data => {
-            return res.json({ status: 200, data: data })
+        }).then(async data => {
+
+            if (data == null) {
+                return res.json({ status: 400, message: "El usuairio no existe" })
+            }
+
+            if (data.usuario.toLowerCase() == "admin") {
+                return res.json({ status: 400, message: "El admin no puede ser actualizado" })
+            }
+
+            await prisma.usuarios.update({
+                where: {
+                    id: parseInt(id)
+                },
+                data: {
+                    nombre: name.toLowerCase(),
+                    usuario: usuario.toLowerCase()
+                }
+            }).then(data => {
+                return res.json({ status: 200, data: data })
+            }).catch(error => {
+                console.log(error);
+                return res.json({ status: 500, message: "Server error" })
+            })
+
         }).catch(error => {
             console.log(error);
             return res.json({ status: 500, message: "Server error" })
         })
+
+
+
 
     },
 
     delete: async (req, res) => {
 
-        const id = req.body.id ?? ""
+        const id = req.params.id ?? ""
 
         if (!isNumeric(id)) {
             return res.json({ status: 400, message: "No pasó la validación" })
         }
 
-        await prisma.usuarios.delete({
+        await prisma.usuarios.findFirst({
             where: {
                 id: parseInt(id)
             }
-        }).then(data => {
-            return res.json({ status: 200, data: data })
+        }).then(async data => {
+
+            if (data.usuario.toLowerCase() == "admin") {
+                return res.json({ status: 400, message: "El admin no puede ser eliminado" })
+            }
+
+            await prisma.usuarios.delete({
+                where: {
+                    id: parseInt(id)
+                }
+            }).then(data => {
+                return res.json({ status: 200, data: data })
+            }).catch(error => {
+                console.log(error);
+                return res.json({ status: 500, message: "Server error" })
+            })
+
         }).catch(error => {
             console.log(error);
             return res.json({ status: 500, message: "Server error" })
         })
-
     },
 }
